refactor: extract Sidebar component to components/shared/Sidebar.tsx
