+++ aloop/cli/dashboard/src/App.tsx
@@ -1,33 +1,13 @@
 import * as view from './AppView';
+import { Sidebar } from './components/shared/Sidebar';
+import { stripAnsi, rgbStr, parseAnsiSegments, renderAnsiToHtml } from './lib/ansi';
+import { isRecord, str, numStr, toSession, formatTime, formatTimeShort, formatSecs, formatDuration, formatDateKey, relativeTime } from './lib/format';
 
+export { Sidebar };
 export const App = view.App;
-export const Sidebar = view.Sidebar;
-export const ActivityPanel = view.ActivityPanel;
-export const DocContent = view.DocContent;
-export const HealthPanel = view.HealthPanel;
-export const SIGNIFICANT_EVENTS = view.SIGNIFICANT_EVENTS;
-export const IMAGE_EXT = view.IMAGE_EXT;
-
-export function stripAnsi(text: string): string { return view.stripAnsi(text); }
-export function rgbStr(r: number, g: number, b: number): string { return view.rgbStr(r, g, b); }
-export function parseAnsiSegments(text: string) { return view.parseAnsiSegments(text); }
-export function renderAnsiToHtml(text: string, opts: { gfm?: boolean; breaks?: boolean } = {}): string {
-  return view.renderAnsiToHtml(text, opts);
-}
-
-export function isRecord(v: unknown): v is Record<string, unknown> { return view.isRecord(v); }
-export function str(source: Record<string, unknown>, keys: string[], fb = ''): string { return view.str(source, keys, fb); }
-export function numStr(source: Record<string, unknown>, keys: string[], fb = '--'): string { return view.numStr(source, keys, fb); }
-export function toSession(source: Record<string, unknown>, fallback: string, isActive: boolean) {
-  return view.toSession(source, fallback, isActive);
-}
-
-export function formatTime(ts: string): string { return view.formatTime(ts); }
-export function formatTimeShort(ts: string): string { return view.formatTimeShort(ts); }
-export function formatSecs(total: number): string { return view.formatSecs(total); }
-export function formatDuration(raw: string): string { return view.formatDuration(raw); }
-export function formatDateKey(ts: string): string { return view.formatDateKey(ts); }
-export function relativeTime(ts: string): string { return view.relativeTime(ts); }
+export { ActivityPanel, DocContent, HealthPanel, SIGNIFICANT_EVENTS, IMAGE_EXT } from './AppView';
+export { stripAnsi, rgbStr, parseAnsiSegments, renderAnsiToHtml };
+export { isRecord, str, numStr, toSession, formatTime, formatTimeShort, formatSecs, formatDuration, formatDateKey, relativeTime };
 
 export function parseLogLine(line: string) { return view.parseLogLine(line); }
 export function isImageArtifact(a: { type: string; path: string; description: string; metadata?: { baseline?: string; diff_percentage?: number } }) {
